requirements.txt
flask
flask-cors
gunicorn
mysql-connector-python
python-dotenv
flask-jwt-extended
bcrypt
